+++ jlib-reflect/src/main/java/org/jlib/reflect/languageelement/corereflection/ReflectionLanguageElementHandler.java
@@ -19,7 +19,7 @@
  *     limitations under the License.
  */
 
-package org.jlib.reflect.languageelement.reflection;
+package org.jlib.reflect.languageelement.corereflection;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
