+++ jlib-reflect/src/test/java/org/jlib/reflect/languageelement/corereflection/ReflectionTestMethods.java
@@ -19,7 +19,7 @@
  *     limitations under the License.
  */
 
-package org.jlib.reflect.languageelement.reflection;
+package org.jlib.reflect.languageelement.corereflection;
 
 @SuppressWarnings({ "EmptyMethod", "SameReturnValue" })
 class ReflectionTestMethods {
